refactor: encapsulated URL of Kakao API into build.properties
